requirements.txt
antlr4-python3-runtime==4.13.2
antlr4-tools==0.2.1
install-jdk==1.1.0
setuptools==78.1.0
wheel==0.45.1

pygame~=2.6.1
numpy~=2.2.5
PyOpenGL~=3.1.9